Chore: Setup Nest.js api template

+++ src/api/tasks/tasks.service.ts
@@ -1,34 +1,84 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  BadRequestException,
+  NotFoundException,
+} from '@nestjs/common';
+import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm/repository/Repository';
+
+import { UsersService } from '../users/users.service';
 
 import { Task } from './entities/task.entity';
-import { CreateTaskDto, UpdateTaskDto } from './dto';
+import { CreateTaskDto } from './dto/create-task.dto';
+import { UpdateTaskDto } from './dto/update-task.dto';
+import { TaskResponseDto } from './dto/task-response.dto';
+import { TaskResponseFactory } from './factories/task-response.factory';
 
 @Injectable()
 export class TasksService {
   constructor(
     @InjectRepository(Task) private tasksRepository: Repository<Task>,
+    private userService: UsersService,
   ) {}
 
-  create(createTaskDto: CreateTaskDto) {
+  async create(createTaskDto: CreateTaskDto): Promise<TaskResponseDto> {
+    const userExists = await this.userService.exists(createTaskDto.userId);
+
+    if (!userExists) {
+      throw new BadRequestException('User does not exist');
+    }
+
     const task = this.tasksRepository.create(createTaskDto);
-    return this.tasksRepository.save(task);
+    const createdTask = await this.tasksRepository.save(task);
+
+    return TaskResponseFactory.for(createdTask);
   }
 
-  findAll() {
-    return this.tasksRepository.find();
+  async findAll(): Promise<TaskResponseDto[]> {
+    const tasks = await this.tasksRepository.find();
+    return tasks.map((task) => TaskResponseFactory.for(task));
   }
 
-  findOne(id: number) {
-    return this.tasksRepository.findOneBy({ id });
+  async findOne(id: number): Promise<TaskResponseDto> {
+    const task = await this.tasksRepository.findOneBy({ id });
+
+    if (!task) {
+      throw new BadRequestException('Task not found');
+    }
+
+    return TaskResponseFactory.for(task);
   }
 
-  update(id: number, updateTaskDto: UpdateTaskDto) {
-    return this.tasksRepository.update(id, updateTaskDto);
+  async update(
+    id: number,
+    updateTaskDto: UpdateTaskDto,
+  ): Promise<TaskResponseDto> {
+    const task = await this.tasksRepository.preload({
+      id,
+      ...updateTaskDto,
+    });
+
+    if (!task) {
+      throw new BadRequestException('Task not found');
+    }
+
+    const updatedTask = await this.tasksRepository.save(task);
+
+    return TaskResponseFactory.for(updatedTask);
   }
 
-  remove(id: number) {
-    return this.tasksRepository.delete(id);
+  async remove(id: number): Promise<void> {
+    const taskExists = await this.tasksRepository.findOneBy({ id });
+
+    if (!taskExists) {
+      throw new NotFoundException('Task not found');
+    }
+
+    await this.tasksRepository.delete(id);
+  }
+
+  async exists(id: number): Promise<boolean> {
+    const task = await this.tasksRepository.findOneBy({ id });
+    return !!task;
   }
 }
